requirements.txt
asgiref==3.8.1
certifi==2025.1.31
Django==5.1.7
django-environ==0.12.0
django-graphql-jwt==0.4.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
promise==2.3
psycopg2==2.9.10
PyJWT==2.10.1
pyproj==3.7.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
utm==0.8.1
